# swagger.yaml
openapi: "3.0.3"
info:
  title: Himnario API
  description: REST API para gestión de himnos. Construida en Go con Gin y arquitectura hexagonal.
  version: "1.0.0"

servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo

# ---------------------------------------------------------------------------
# Headers globales requeridos en TODOS los endpoints
# ---------------------------------------------------------------------------
components:
  parameters:
    XClient:
      name: x-client
      in: header
      required: true
      schema:
        type: string
        enum: [mbh, front-end-himnary]
      description: Identifica al cliente consumidor de la API.
    Country:
      name: country
      in: header
      required: true
      schema:
        type: string
        enum: [CL]
      description: País del cliente. Único valor aceptado es CL.
    EventId:
      name: event-id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Identificador único de la solicitud. Cualquier string no vacío.
    Accept:
      name: Accept
      in: header
      required: true
      schema:
        type: string
        enum: ["application/json", "*/*"]

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtenido desde POST /user/authentication. Expira en 60 minutos.

  schemas:
    # --- Wrappers de respuesta ---
    ResponseOk:
      type: object
      properties:
        responseOk:
          type: object
          properties:
            statusCode:
              type: integer
              example: 200
            message:
              type: string
            data:
              description: Payload de la respuesta. Puede ser objeto, array o null.

    ResponseError:
      type: object
      properties:
        responseError:
          type: object
          properties:
            statusCode:
              type: integer
            error:
              type: string
              description: Mensaje descriptivo del error.
            data:
              type: object
              additionalProperties:
                type: string
              description: Mapa de campos con sus mensajes de error.

    # --- Dominio ---
    HymnSummary:
      type: object
      description: Himno sin contenido, usado en listados.
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Himno de ejemplo"
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"

    HymnFull:
      type: object
      description: Himno completo incluyendo contenido JSONB.
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Himno de ejemplo"
        content:
          type: object
          description: Estructura libre almacenada como JSONB en PostgreSQL.
          example:
            verses:
              - number: 1
                lines: ["Línea uno", "Línea dos"]
            chorus: "Estribillo del himno"
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"

    UserInBody:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          description: Debe coincidir exactamente con el email del JWT.
          example: "editor@ejemplo.com"
        name:
          type: string
          description: Debe coincidir exactamente con el nombre del JWT.
          example: "Nombre Editor"

    # --- Requests ---
    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: "usuario@ejemplo.com"
        password:
          type: string
          minLength: 8
          example: "minimo8caracteres"
        name:
          type: string
          example: "Nombre Apellido"
        role:
          type: string
          enum: [admin, editor, viewer]
          example: "viewer"

    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "usuario@ejemplo.com"
        password:
          type: string
          minLength: 8
          example: "minimo8caracteres"

    CreateHymnRequest:
      type: object
      required: [title, content, user]
      properties:
        title:
          type: string
          description: Debe ser único (case-insensitive).
          example: "Título del himno"
        content:
          type: object
          description: Estructura libre del himno (se almacena como JSONB).
          example:
            verses:
              - number: 1
                lines: ["Línea uno", "Línea dos"]
            chorus: "Estribillo del himno"
        user:
          $ref: "#/components/schemas/UserInBody"

    EditHymnRequest:
      type: object
      required: [title, content, user]
      properties:
        title:
          type: string
          description: Nuevo título del himno (case-insensitive único).
          example: "Nuevo título del himno"
        content:
          type: object
          description: Nuevo contenido del himno (estructura libre, JSONB).
          example:
            verses:
              - number: 1
                lines: ["Línea actualizada"]
            chorus: "Estribillo actualizado"
        user:
          $ref: "#/components/schemas/UserInBody"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:

  /health:
    get:
      summary: Health check
      description: Verifica que el servidor está activo.
      tags: [Sistema]
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
      responses:
        "200":
          description: Servidor activo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "ok"
        "403":
          description: Ruta no permitida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"

  /user/create:
    post:
      summary: Registrar usuario
      description: Crea una nueva cuenta de usuario.
      tags: [Usuarios]
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: Usuario creado exitosamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "User created successfully"
                  data: null
        "400":
          description: Body o headers inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 400
                  error: "Invalid request body"
                  data:
                    password: "debe tener mínimo 8 caracteres"
        "409":
          description: El email ya está registrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 409
                  error: "User creation failed"
                  data:
                    email: "email already exists"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"

  /user/authentication:
    post:
      summary: Autenticar usuario
      description: Retorna un JWT válido por 60 minutos y los datos del usuario autenticado.
      tags: [Usuarios]
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Autenticación exitosa. Retorna JWT y datos de usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "Authentication successful"
                  data:
                    email: "usuario@ejemplo.com"
                    name: "Nombre Apellido"
                    role: "viewer"
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Body o headers inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "401":
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 401
                  error: "Authentication failed"
                  data:
                    authentication: "invalid credentials"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"

  /hymn/create:
    post:
      summary: Crear himno
      description: |
        Crea un nuevo himno. Requiere JWT con rol `admin` o `editor`.
        El email y nombre del campo `user` deben coincidir exactamente con los del token.
      tags: [Himnos]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHymnRequest"
      responses:
        "200":
          description: Himno creado exitosamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "Hymn created successfully"
                  data: null
        "400":
          description: Body o headers inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 400
                  error: "Invalid request body"
                  data:
                    title: "campo requerido"
        "401":
          description: Token inválido, expirado o datos de usuario no coinciden con el token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 401
                  error: "Unauthorized"
                  data:
                    error: "invalid or expired token"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 500
                  error: "Hymn creation failed"
                  data:
                    error: "internal server error"

  /hymn/all:
    get:
      summary: Listar todos los himnos
      description: Retorna todos los himnos ordenados por título. No incluye el contenido.
      tags: [Himnos]
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
      responses:
        "200":
          description: Lista de himnos (sin contenido).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "Get all hymns successfully"
                  data:
                    - id: 1
                      title: "Himno de ejemplo"
                      created_at: "2025-01-15T10:30:00Z"
                    - id: 2
                      title: "Otro himno"
                      created_at: "2025-02-01T08:00:00Z"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"

  /hymn/{id}:
    get:
      summary: Obtener himno por ID
      description: Retorna un himno completo incluyendo su contenido JSONB.
      tags: [Himnos]
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del himno.
          example: 1
      responses:
        "200":
          description: Himno completo con contenido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "Get hymn successfully"
                  data:
                    id: 1
                    title: "Himno de ejemplo"
                    content:
                      verses:
                        - number: 1
                          lines: ["Línea uno", "Línea dos"]
                      chorus: "Estribillo del himno"
                    created_at: "2025-01-15T10:30:00Z"
        "400":
          description: El ID no es un número entero válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 400
                  error: "Invalid hymn ID"
                  data:
                    error: "strconv.Atoi: parsing \"abc\": invalid syntax"
        "404":
          description: Himno no encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 404
                  error: "Hymn not found"
                  data:
                    error: "hymn with given ID does not exist"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"

    put:
      summary: Editar himno por ID
      description: |
        Actualiza el título y contenido de un himno existente. Llama al stored procedure
        `sp_update_hymn` que verifica la existencia del himno antes de actualizar.

        Requiere JWT con rol `admin` o `editor`.
        El email y nombre del campo `user` deben coincidir exactamente con los del token.
      tags: [Himnos]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XClient"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Accept"
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del himno a editar.
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditHymnRequest"
            example:
              title: "Nuevo título del himno"
              content:
                verses:
                  - number: 1
                    lines: ["Línea actualizada"]
                chorus: "Estribillo actualizado"
              user:
                email: "editor@ejemplo.com"
                name: "Nombre Editor"
      responses:
        "200":
          description: Himno editado exitosamente. Retorna el himno con los datos actualizados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOk"
              example:
                responseOk:
                  statusCode: 200
                  message: "edit hymn successfully"
                  data:
                    id: 1
                    title: "Nuevo título del himno"
                    content:
                      verses:
                        - number: 1
                          lines: ["Línea actualizada"]
                      chorus: "Estribillo actualizado"
                    created_at: "2025-01-15T10:30:00Z"
        "400":
          description: ID inválido o body con campos faltantes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              examples:
                id_invalido:
                  summary: ID no es un entero
                  value:
                    responseError:
                      statusCode: 400
                      error: "Invalid hymn ID"
                      data:
                        error: "strconv.Atoi: parsing \"abc\": invalid syntax"
                body_invalido:
                  summary: Campos requeridos faltantes
                  value:
                    responseError:
                      statusCode: 400
                      error: "Invalid request body"
                      data:
                        Title: "Key: 'EditHymn.Title' Error:Field validation for 'Title' failed on the 'required' tag"
        "401":
          description: |
            Token ausente, malformado, expirado, o el email/nombre del body
            no coincide con los del token, o el rol no es admin/editor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              examples:
                token_invalido:
                  summary: Token inválido o expirado
                  value:
                    responseError:
                      statusCode: 401
                      error: "Unauthorized"
                      data:
                        error: "invalid or expired token"
                token_ausente:
                  summary: Header Authorization ausente o malformado
                  value:
                    responseError:
                      statusCode: 401
                      error: "Unauthorized"
                      data:
                        error: "authorization header missing or malformed"
                usuario_no_coincide:
                  summary: Email/nombre no coincide con el token
                  value:
                    responseError:
                      statusCode: 401
                      error: "Unauthorized"
                      data:
                        error: "unauthorized"
        "404":
          description: Himno no encontrado con el ID dado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 404
                  error: "Hymn not found"
                  data:
                    error: "hymn with given ID does not exist"
        "500":
          description: Error interno del servidor al intentar editar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
              example:
                responseError:
                  statusCode: 500
                  error: "Failed to edit hymn"
                  data:
                    error: "internal server error"

# ---------------------------------------------------------------------------
# Tags (orden en la UI)
# ---------------------------------------------------------------------------
tags:
  - name: Sistema
    description: Endpoints de infraestructura.
  - name: Usuarios
    description: Registro y autenticación de usuarios.
  - name: Himnos
    description: Gestión del catálogo de himnos.